backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Stephen Eaton", email = "seaton@strobotics.com.au" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "email-validator>=2.2.0",
    "faker>=37.6.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-jose>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[project.scripts]
portfolio-backend = "portfolio_backend:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.18.1",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["migrations"]
